requirements.txt
pandas==1.5.3
scikit-learn==1.3.2
xgboost==1.7.6
mlflow==2.10.1
psycopg2-binary==2.9.9
sqlalchemy==1.4.50
apache-airflow-providers-postgres
